move non-faerie-specific parts of .eh_frame generation to cranelift-module
