+++ cranelift-module/Cargo.toml
@@ -13,6 +13,7 @@ edition = "2018"
 [dependencies]
 cranelift-codegen = { path = "../cranelift-codegen", version = "0.54.0", default-features = false }
 cranelift-entity = { path = "../cranelift-entity", version = "0.54.0" }
+gimli = "0.19.0"
 hashbrown = { version = "0.6", optional = true }
 log = { version = "0.4.6", default-features = false }
 thiserror = "1.0.4"
